requirements.txt
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
exceptiongroup==1.3.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
natsort==8.4.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.1
polars-runtime-32==1.35.1
psutil==7.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.21
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
ultralytics==8.3.222
ultralytics-thop==2.0.18
urllib3==2.5.0
